Move Texture2DExtensions.GetFloodFillPixels() into new FloodFill class
